涉及单表操作的大部分CRUD操作已完成； 接下来要做的是：完成和多表查询相关的工作；

+++ src/main/java/com/zhannicholas/cpwms/service/impl/CustomerServiceImpl.java
@@ -3,11 +3,15 @@ package com.zhannicholas.cpwms.service.impl;
 import com.zhannicholas.cpwms.domain.model.Customer;
 import com.zhannicholas.cpwms.domain.repository.CustomerRepository;
 import com.zhannicholas.cpwms.service.CustomerService;
+import com.zhannicholas.cpwms.util.ToMapUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Map;
 
 @Service
 @Transactional
@@ -19,28 +23,65 @@ public class CustomerServiceImpl implements CustomerService {
         this.customerRepository = customerRepository;
     }
 
+
     @Override
-    public List<Customer> findAll() {
-        return customerRepository.findAll();
+    public Map<String, Object> findAll(Pageable pageable) {
+        Page<Customer> customerPage = customerRepository.findAll(pageable);
+        return ToMapUtil.fromAPage(customerPage);
     }
 
     @Override
-    public void save(Customer customer) {
-        customerRepository.save(customer);
+    public boolean save(Customer customer) {
+        if(isValidCustomer(customer)){
+            customerRepository.save(customer);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public Customer findOne(int customerId) {
-        return customerRepository.findByCustomerId(customerId);
+    public Map<String, Object> findOne(int customerId) {
+        Customer customer = customerRepository.findByCustomerId(customerId);
+        return ToMapUtil.fromAInstance(customer);
     }
 
     @Override
-    public List<Customer> findByCustomerName(String customerName) {
-        return customerRepository.findByCustomerNameContaining(customerName);
+    public Map<String, Object> findByCustomerNameContaining(String customerName, Pageable pageable) {
+        Page<Customer> customerPage = customerRepository.findByCustomerNameContaining(customerName, pageable);
+        return ToMapUtil.fromAPage(customerPage);
     }
 
     @Override
-    public void delete(int customerId) {
-        customerRepository.deleteById(customerId);
+    public boolean delete(int customerId) {
+        if(isValidSupplierId(customerId)){
+            customerRepository.deleteById(customerId);
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * 检查客户信息是否满足要求
+     * @param customer  客户
+     * @return  若客户信息满足要求则返回true，否则返回false
+     */
+    private boolean isValidCustomer(Customer customer){
+        if(customer != null){
+            return customer.getCustomerName() != null &&
+                    customer.getCustomerPerson() != null &&
+                    customer.getCustomerTel() != null &&
+                    customer.getCustomerEmail() != null &&
+                    customer.getCustomerAddress() != null;
+        }
+        return false;
+    }
+
+    /**
+     * 检查 customer 是否合法
+     * @param customerId    客户id
+     * @return  若客户id合法则返回true,否则返回false
+     */
+    private boolean isValidSupplierId(int customerId){
+        return customerRepository.findByCustomerId(customerId) != null;
     }
 }
